Implemented model sway animations for weapons while aiming down the sights.
Previously, when aiming down the weapon sights, the gun model would stay completely stiff regardless of the movements the player was making. Now, a lovely inertia sway effect will show, similar to the effect on the player’s whole upper body when they adjusted their aim. I had to remake it and provide weapon-specific values, because the player’s upper body rotation would throw off the weapon’s sights, and therefore the player’s aim.
I had to update my weapon’s DetaPosition and DeltaLookRotation code, so they could be used on other functions. Previously, the last frame position and rotation values used to calculate the delta values were changed at the end of LateUpdate(), after the player’s animations were performed. But since these values were needed for the gun ADS animations, which were lower on the hierarchy and therefore performed later in the execution order, the last frame values were already updated and so to the ADS code no change would exist. I solved this by creating an IEnumerator in the MovementController class that activates when the script is enabled and persists until disabled. This script utilises a while loop that yields for WaitForEndOfFrame before updating the values, waiting until after even LateUpdate(). This allows all cosmetic functions running in LateUpdate() in any class to utilise these values properly.
Meanwhile, in my GunADS script, after the base gun rotation is specified so it lines up with the player’s aim direction, an extra rotation is applied. This rotation is very similar to the MovementController’s TorsoSway() function, except the euler angles are modified with Vector3.SmoothDamp to eliminate jittering. In MovementController, the SmoothDamp function is directly used on the torso’s position and rotation, but that didn’t work with the player’s guns because it would affect the base rotation as well, and cause the gun model to lag too much and take far too long to rotate into the correct position. By calculating the base value without any damping, then applying the damped sway on top of that, the rotation is damped slightly while remaining within hard limits and still feeling snappy.
Now that this rotation was applied, the gun’s position needed to be updated so that the player’s reticle would always line up with the direction the player was aiming in. The previous method to do this was to shift the position so a child of the gun model, an empty transform representing the player’s relative head orientation, would line up with the player’s head/upper body transform, from which shots were calculated. But since this position was some distance out from the actual point the gun needed to rotate around, I had to replace this value with a reticle axis, at the same X and Y positions but moved forward to the actual point on the gun the player is supposed to focus on.
Since this change would result in the gun model shifting inside the player, I changed the position by a Vector3 equivalent to forward in the player’s aim direction, multiplied by a distance magnitude. This value had to be calculated on the fly instead of preserving the relative head orientation transform, as that would shift around with the cosmetic rotation due to being a child of the weapon model itself.
This cosmetic system required a lot of math, and may be difficult to grasp, but I am incredibly pleased with the way it looks. My shooter is starting to look more and more polished and professional.