Fixed gun firing animations to work with ADS
I tweaked the way the ADS position calculation and gun firing animations work, so they still work when the weapon is aiming down the sights. Each gun’s model now has two transforms for calculating position – a model orientation transform that determines the gun’s regular position, and the model’s actual transform which is a child of the first transform. Using this method, I can have the model orientation transform update based on factors like switching, reloading or ADS functionality, while the actual model transform performs its firing animation and moves relative to the orientation transform. This way, the firing animations will function the same way if the player is hipfiring or using ADS.
Previously, the animations would not trigger if the weapon was aiming down sights, as the movement code in LateUpdate() was being overridden by similar code in LateUpdate() for shifting the gun firing position between hipfiring and aiming down sights. Animations can still be implemented this way to deliberately disable the firing animation while using ADS, in case a gun’s recoil animation is too severe and looks awkward while ADS is active. This is a placeholder animation, however, and there will need to be a better long-term solution for having different animations play based on if the player is hipfiring or using ADS.

I also added an extra animation to the shotgun, that visibly pumps the forend.