Implemented graphic and description functionality for menus
My menu options have had variables for graphics and descriptions, but now that information will actually appear.
I added a reference in the MenuInteractable class to its parent component, that it obtains in Awake() using GetComponent<>(). I then implemented a function named RefreshSelectionInfo() that automatically updates the menu’s selection graphic and description to display the MenuInteractable’s info. Finally, I then changed the MenuInteractable class to include the ISelectHandler and IPointerEnterHandler interfaces, and used their functions to run RefreshSelectionInfo() automatically when the interactable is selected or moused over by an EventSystem.
I also added code to automatically refresh this data when a menu is switched to. I added a ‘firstSelection’ variable to the MenuWindow class. When that window is switched to, it will signal the EventSystem to automatically assign ‘firstSelection’ as the currently selected object, ensuring that a function can be selected without having to mouse over it. This will be important for players who prefer to use a controller or keyboard to navigate menus.
I also removed some unnecessary code in the MenuWindow class for handling this functionality, as the way I actually went with had the code concentrated in other classes.