Scope shader with mask
The problem with the previous way I implemented the scope effects is that there was no way to have a mask and a custom material. If I used Unity’s default UI Mask features, it would override any custom material on the sight picture renderer, and assigning the RenderTexture to ‘no material’ would result in all UI images with no specific material assigned output the RenderTexture display. If I assigned a custom material to an image then overrode the texture with my RenderTexture, that material wouldn’t have any mask functionality and the image would always be a square, regardless of the shape of the weapon’s scope. To solve this, I had to create a simple shader for my scope, that utilised two textures – a main texture that gets overridden by the RenderTexture representing the scoped view, and a mask texture that cuts out the unnecessary sections. This shader is not perfect – it does not allow any kind of light or reflections that would make sense for a physical object, and I was informed that the code I used would not function well on mobile devices. But considering my lack of experience with shaders, I was very pleased with the result.