Camera shake when shooting
I created a simple camera shake function, to add extra physicality to certain game actions. Currently, it is implemented in the firing animations of the player’s weapons.
The code consists of a MonoBehaviour featuring cosmetic stats, as well as a maximum range and falloff curve. This script contains an IEnumerator that references a player’s MovementController class, with an individual intensity multiplier. This function applies camera shaking by obtaining two noise values and rotating the player’s camera relative to their current aim direction, for a short period of time using a curve to control intensity. A public Play() function is also present, which uses FindObjectsOfType to find the player (or potentially any other players in case I add some kind of multiplayer functionality) and trigger the IEnumerator, using the distance and falloff curve to determine the proper intensity. This Play() function exists to make it easy for me to include a Camera Shake function on any object in the scene and have it affect the player without having to directly link them.
There are ways I could improve this function. Currently, it only really works via area of effect, and there are certain actions that can jolt a player around that would only affect a specific point, such as being shot or colliding with an object. However, it presently works well enough that I am willing to put it aside and work on other projects.