Cleaned up several weapon functions to require less unnecessary assignment of parameters
Most of the scripts required for my weapon systems required references to other scripts higher in the hierarchy. E.g. a gun stat class such as the fire controller requires a reference to its RangedAttack weapon mode, which requires a reference to the weapon handler, and so on. To most of these scripts, I added permanent references for the required classes, which are then assigned during OnEnable(), Awake() or inside one of the loops for input control. This will make my code much cleaner and easier to understand and maintain.