Added crouch toggle, and improved code for it and ADS toggle
I created a static bool named SetPlayerAbilityState(), in the custom input class, that makes it easy to change a player function that can be either toggled or held. In this game, it’s used for crouching and ADS functionality. I then updated my player’s crouching code to include it. (I also modified the ADS code to include it, but I accidentally uploaded that in a different push.)
I also moved the ButtonInput struct into the CustomInput class, and renamed it just to Button, and reworked the player’s jumping code to use it.