Gun ADS sway code now supports swaying in the opposite direction
Formerly, the if statement that activates the gun swaying code only returned true if the lookSwayDegrees value was greater than zero. I changed this to instead check if the value is simply not equal to zero, because this meant that if I wanted a gun to swing in the opposite direction for some reason, I wouldn't be able to.