ADS functionality is interrupted by and cannot engage while the player is reloading.
Previously, the player could still activate ADS while in the middle of a reload animation, messing up the reload animations. This is now fixed, although it can cause a slightly different problem where as soon as the reload is finished the gun will automatically activate the ADS animation without waiting for the weapon to return to the default position.