Gun accuracy deviation from recoil
In addition to the camera shake and the animation of the player’s gun jolting back with recoil, I recently implemented a mechanic so when firing the gun, the player’s aim will be thrown off as well.
In the GunGeneralStats class, which determines the effects of a single shot of one of the player’s weapons, I added stats for recoil magnitude (how many degrees the player’s aim is thrown off), deviation angle (in what possible range of directions), and time (how long the gun takes to move that far). The first two stats are used to calculate a Vector2 representing a direction and magnitude, which along with the time are fed into an IEnumerator in the MovementController class, named RotateAimOverTime(), which increments the player’s aim by said Vector2, over said time frame.
This mechanic has done a lot to improve the feel of the combat, both by making guns feel more cathartic and by adding an extra layer of skill in controlling the weapon’s recoil for accurate follow-up shots.