Cleaned up player aiming control code
Since the player’s current camera input is obtained via a property, I decided to streamline my code by taking the code that multiplies by sensitivity and including it in the same property. Since rotating the player’s aim using raw angle data is done in the separate RotateAim() function, so it can be used by other processes that alter the player’s aim direction, the sensitivity calculations were included in a second function. Since this code is now in the original property that obtains the player’s aim values, I have been able to remove this function for cleaner and simpler code.
I also tweaked the code to add dedicated bools for inverting the X and Y axes. My original idea was to simply invert the aim if the sensitivity values were negative, so I could have a single aim slider in the options menu that also allowed negative values, but when this would allow the player from putting their aim sensitivity at zero, I realised the best solution was just to limit the slider from going to zero, and handle the inversion by using bools to invert the input values.