Modified placeholder scope to look slightly better
I modelled a simple tube in Autodesk Maya and imported it to use as the model. Since the scope now has an actual interior, I was able to add a pleasant rim to the scope lens by moving the sight picture overlay further forward into the tube.
I made this change because I was experimenting with different scope effects for improved visuals. I eventually decided I had more important things to work on, but I left this particular change because it looked much nicer.