Smoother recoil aim deviation
I made a slightly tweaked version of the recoil function that uses an animation curve to make the camera move smoothly with acceleration and deceleration. Since the nature of the recoil function and how multiple circumstances can influence it at the same time means we can’t use lerping, I accomplished this by creating a value in the IEnumerator that determines the curve value on the previous frame. Every frame, the main timer ticks up at a linear rate and is used to evaluate a value from the curve. This is checked against the previous frame value to produce a value essentially like Time.deltaTime, but warped by the curve. This value is then used to apply the recoil at a rate over time that is influenced by the curve.