Fixed jittering with movement controller torso tilt and drag animations
I realised the cause was the way I was trying to infer the total velocity. My original method was to calculate a delta movement value, then divide it by Time.deltaTime to get a rough approximation of the speed the player was moving that frame, accounting for both rigidbody velocity and the player’s current movement direction. But after checking the raw value this was producing, I realised the distance was causing an unusual jittering effect.
I then had to create a TotalVelocity property. This combines the rigidbody’s velocity with a new Vector3 named movementVelocity, which registers the speed in which the player is walking using Rigidbody.MovePosition. I then had to tweak my movement code so this value is updated in Update() based on the player’s movement controls, then referenced for actual movement in FixedUpdate(), instead of controls being directly registered and calculated in FixedUpdate(). This code is slightly more complicated, but gives me a much smoother and less jittery total velocity value, leading to smoother drag and tilt animations.