Fixed movement controller cosmetic torso and gun sway to be more consistent with framerate
I performed a similar fix to with the torso drag and tilting, by removing the delta rotation values and replacing them with a more consistent method. I noticed, while I was updating the tilt and drag functions, that the sway function for the player's torso (and their gun while in ADS) relied on raw delta values without any multiplication or division to account for a varying framerate. I changed this accordingly, so the animations should now act consistently regardless of framerate.