Main menu tweaks – selection descriptions, renamed title, disabled nonfunctional options and a nice background
Renamed the title to actually read ‘Main Menu’.
Added a grey skybox and a background visual of some of the player’s guns, as a nice placeholder background graphic to have instead of the default Unity skybox look.
Added a description to the ‘Load Game’ button.
Disabled all the menu options that don’t have any functionality assigned to them. They still appear for visual clarity but are highlighted red and cannot be interacted with.