Updated MenuWindow’s Start() function to prevent unnecessary extra switching
During Start(), each MenuWindow object would run SwitchWindow(this); to switch to itself. Since I don’t want it to occur with every one, which would result in the first selected window being based on the hierarchy, I added a check so the menu only runs the initialising switch function if it is the root.