MenuWindow’s SwitchWindow() function now updates interactables to prevent conflict with the currently open window
When a new window is switched to, when a window needs to be enabled or disabled, a new function named SetActiveState(bool active) is run. As well as altering the Canvas Group’s alpha value to determine visibility (like what was happening in the previous iteration), it also updates its ‘interactable’ bool to prevent the EventSystem from transitioning to menu objects that are supposed to be hidden and non-interactable, and its ‘blocksRaycasts’ bool to ensure menu objects in windows lower in the hierarchy (and therefore in front) do not block interaction with other menus higher in the hierarchy but currently enabled.