CustomInput.Button struct now uses keycodes, has controller options
I removed the Key struct, and changed the Button struct to include keycodes instead of strings for Input.GetButton(). I also created a simple enum for controller buttons, as an easy way to determine controller inputs without messing around with the joystick options in the KeyCode enum, which are confusingly named and have multiple variants for each different controller.
I think I am going to have to convert to Unity’s new input system, to have the level of functionality I want.