Started implementing GUIButtonPrompt class for GUI elements that show the specific button needed to perform an action
I wanted to implement this as a simple method of player training. When the Refresh() function is run, using a button struct as a parameter, it will check the assigned keycode to update an Image and Text object to display a key and the correct button name. If a controller button prompt is shown, the text will be empty and the image will display one of a list of graphics corresponding to each button.

This functionality is not fully tested yet, but it is something that needs to be implemented at some point.