Interact function
I implemented a simple system for allowing the player to interact with objects in a way other than shooting them.
I started by creating an ‘Interactable’ MonoBehaviour that attaches to an object with a collider, and includes a descriptive action and a UnityEvent, using a Player as a parameter. I also created an InteractFunction class, which attaches to the player, and is used for interacting with things. This works by continuously checking the direction the player is looking in using a short raycast, to find whatever interactable object the player is currently looking at. If one is found, the player can press an interact button to invoke the Interactable’s UnityEvent.
It’s a simple and functional interaction function like what many first-person shooters have implemented, and although I will need to make it look slightly prettier, it works well enough for my current purposes.