AmmoRefill class
Created a new class to allow the player to refill their ammunition. It works in conjunction with the Interactable class, with a function using a Player as a parameter, that is triggered by the Interactable. Since I wanted functionality to only take small amounts of health, I slightly tweaked the AmmunitionInventory's Spend function to output a remainder variable, so the code will know how much is leftover after the player takes some.