Cooldown and improved visual clarity for interaction function
Added an ‘active’ bool so interactables can be disabled.
Added a cooldown feature, so interactions cannot be spammed. This is activated using a UnityEvent, in case I want an interactable object that doesn’t have a cooldown.
Added separate messages that display depending on if it can be interacted with.
Added a progress bar that shows how long the cooldown has left to go.
To accommodate the new GUI features, I moved all the GUI code from InteractFunction to a new class called InteractGUIPrompt, which contains all the visual elements and code – a Refresh() function that updates the name and button prompt, and a LateUpdate() call which updates the progress bar and descriptive text.
Renamed SendTestMessage() to PrintDebugLog() for improved clarity.