Rifle animations now rely on Unity’s Animator system
My rifle now uses proper animations and does not rely on placeholder lerping code.
To trigger the weapon animation variables without relying on hard-coding things into the weapon class, I created a new MonoBehaviour named WeaponAnimator. This class references the weapon and the Animator, and uses weapon variables to update the appropriate animator variables. It also makes references to various UnityEvents in the weapons to add listeners for functions that aren’t updated continuously, like trigger variables for firing and switching weapon modes.
I also had to slightly tweak the GunMagazine class to make it work better with actual animations. There are now two UnityEvents that trigger each time a new round is loaded into a magazine – onIncrementStart and onIncrementEnd, making it easier to run functions both when the animation starts and when it finishes. The former has a listener added to it by the WeaponAnimator, to trigger the next stage of the reloading process.
I am very pleased with this. Not only did I fix an annoying bug where the SimpleLerpAnimator would mess up the player’s ADS animations (which are still performed via lerping code due to the extra math required to ensure the player’s reticle is in the appropriate part of the screen), but now I have made a framework that will make it much easier to slot in proper animations later down the line.