Reworked shotgun to use Animator system instead of placeholder lerping
I gave the shotgun a similar overhaul to what I did to the assault rifle. Now, it uses the Animator system and a proper state machine, setting the system up to allow me to implement proper animations later on. This also fixed a bug where my placeholder lerping animation code interfered with my intended ADS code.

I also found and fixed a bug, which I’ll document here because I thought it was an interesting learning experience.
I ran into a weird animator bug where, when firing the player's shotgun, after switching from the shotgun to another weapon then back, the shotgun will 'disappear' when firing. I checked the Inspector and it's not actually disappearing, it's teleporting to the 'holstered' position. The animation still plays, because the transform used for orienting the weapon for stuff like hipfiring, holstering and ADS is separate from the actual model transform (so the animation plays the same while using ADS).
I think that when an animation doesn't specify values for a particular variable, it defaults to whatever's in the entry state. I didn't specify the orientation transform values during the firing animation because it wasn't necessary, but I think that caused the orientation to default to the entry animation, i.e. the 'holstered' position where it's completely out of sight. I fixed this by obtaining the default idle orientation keyframes from the idle animation and sticking them in the firing animation. It gets overridden by my lerping code for ADS so it doesn't cause problems there, but it prevented the model snapping to the default position because values were still assigned for the duration of the animation.