Added sounds to shotgun firing and reloading, and updated WeaponAnimator to more easily facilitate weapon sounds
I added extra sounds to the firing and reloading of the shotgun – a pump that plays while firing, a noise for each shell being inserted into the chamber, and a final pump when the player is done reloading.
The reloading noises were added via the UnityEvents in the gun’s reload sequence, but the firing animation had to be added via an animation event. Since these are finicky to use and will only run functions from the same object as the Animator, with one parameter each, I had to code a better modular solution. In the WeaponAnimator class, I added a public AudioClip array and a reference to an AudioSource. I then created a public function named PlaySoundFromArray(int index), with a single index used to specify a clip to play. I could then easily reference PlaySoundFromArray() in any of my animation events, using the index to determine the sound I want. It’s not a perfect solution, but it works and was simple to implement, so I won’t lose very much time if I decide I need to change it.
These changes have made my shotgun far more satisfying to use.