Imported options menu code
I have started work on creating an options menu, for the purpose of player accessibility. The core functionality was relatively easy to set up, as I was able to import an abstract OptionsMenu class I had worked on in another project, that I had put some effort into being very easy to develop variants of.