Fixed bug where field of view is not changed in-game after updating in options menu.
Previously, the change would only take effect once the player activated ADS on one of their weapons.
I came up with a slightly roundabout change, but one which adds extra functionality. The camera value is instead changed whenever the FOV slider is changed or refreshed, allowing the player to preview the change inside the options menu.
I also realised that having this code change the camera FOV immediately would cause problems where the magnified FOV would not change accordingly, if the player was using ADS but had finished transitioning into it. The code for lerping the FOV in this way only ran while transitioning, but since it isn’t very costly I changed it to run all the time, to fix this problem.