Tweaked slider to have input field, for improved clarity on current value and to allow changing via typing
I created a ComplexSlider class attached to my slider prefab, with a reference to a slider and an input field. The input field is updated to only accept whole or decimal numbers, depending on if the slider is set to accept those, and listeners are added so if either the slider or the input field is updated, the other will change to match. While not working 100% perfectly, it still works well in allowing greater variety of inputs, and for the player to easily tell the slider’s current value.