Fixed bugs preventing slider/text combo input from updating properly, or from the player being able to input the correct value.
If the player types in a value outside the acceptable range, the code to update the slider will clamp the value and update the text box accordingly. The problem was, if the player was in the middle of typing a multi-digit number and the minimum value was also a multi-digit number, the input field would automatically update and deselect itself partway through typing. I fixed this by only having the update happen once the player confirms their input, rather than whenever they type in a new digit.
I also added an OnEnable() function that updates the input field, since it was not updating properly when the player loaded the options menu.
I also removed a line of clamping code, which was redundant because the slider would clamp it automatically.