Simplified resolution and refresh rate adjustment code using a slider value
This code is turning out to be a real hassle.
Changed refresh rate option from a dropdown that updates whenever the player adjusts the resolution, to a slider representing a target value. The minimum and maximum values for the slider are determined by finding the minimum and maximum possible refresh rates, as shown in Screen.resolutions. This option is disabled if the fullscreen mode is not set to ‘Exclusive Fullscreen’.
I made this change after finding out some quirks of Unity’s code. Firstly, when assigning the refresh rate, you can input any number you like and Unity will automatically find the closest ideal value. And secondly, adjusting the refresh rate only works if the game is in exclusive fullscreen mode.
By making these changes, I was able to greatly simplify the options code for the resolution and refresh rate. My code has been causing a lot of problems, so I’m carving it down to a version that does what it’s supposed to in a way that’s as unlikely as possible to cause glitches.