HUD script for updating weapon info in selector
I created the WeaponSelectorHUD class that contains references to images and text boxes to display a weapon name, mode name, ammo counter and ammo icon. It populates these using its Refresh() function, which uses a RangedAttack class as a parameter.
In the WeaponHandler class, I added a reference to an instance of this, and added a listener to the weapon selector that automatically runs Refresh() in the WeaponSelectorHUD using the currently selected RangedAttack mode.