Simplified weapon controls and update loops for improved simplicity and compatibility with new input system
Now that my control system is moving away from using Update() to register input, I realised my weapon code was now needlessly complicated. I resolved to fix this, to make the code more reliable in future.
Completely removed InputLoop() functions from the WeaponMode() class and other weapon stat classes that ran from it. This was intended to handle weapon functions only when the WeaponHandler was ready to resume normal weapon operation (e.g. when not switching or otherwise in the middle of another action), but became redundant when I had to add a LateUpdate() function to my GunADS class, defeating the point.
Added Update() functions to GunADS and GunMagazine for functionality that was originally in InputLoop().
Added Initialise() functions to GunADS and GunMagazine to set up code such as establishing references to the current weapon mode.
Added abstract OnSwitchTo() function in WeaponMode, used to run Initialise() functions in that mode’s assigned stat classes when that mode is switched to.
Added abstract OnSwitchFrom() function in WeaponMode, for disabling stat classes when the player switches to a different mode.
Added abstract OnPrimaryInput(), OnSecondaryInput() and OnTertiaryInput() functions to WeaponMode, which are activated from WeaponHandler using the events for weapon controls. These functions exist so I can easily add different weapon functionality based on the kind of weapon mode, and allow button presses to directly activate attacks and weapon functions without any need for update loops listening for inputs.