Fixed glitches with handgun firing mode switching bugs
Turns out the animation glitches were due to some quirks with the Animator Controller, with certain values not being assigned in an animation clip and defaulting to the values in the entry state. I solved this by adding a second layer for firing and weapon switching animations to work on, so they run separately on top of whatever changes are made in the base layer.