Updated GUIButtonPrompt to work with new input system (untested)
Added fields for an InputAction named ‘assignedInput’ and a PlayerInput named ‘player’. These are assigned using the AssignAction() function.
If the prompt object is active and an InputAction and PlayerInput are assigned, an Update() function will run, that checks the current input device in PlayerInput to display the binding for said device.
Added a series of Sprite fields to reference different keyboard, mouse and controller buttons
Added a Sprite array property that easily collates these sprites (except for the keyboard sprite since it acts as a ‘default’ if a dedicated sprite isn’t available)
Added a readonly string list for corresponding input mapping paths
An individual binding is displayed using a Refresh() function. This obtains the binding’s path and checks it against the ones in the readonly list. If one is found, it uses the corresponding index to display the correct button sprite. If not, the default keyboard key sprite is displayed, and the corresponding text object will read out the button name, formatted to remove the rest of the filepath. Currently, it is not formatted to add capital letters or line breaks, but this shouldn’t be too difficult to add in future.