Fixed first selections in gameplay and video options menu, so the player is not stuck when using a controller
If the first selection in a MenuWindow object is set to an object that is disabled or non-interactable, and the player is not using a pointer-based input for navigating menus, they will be unable to navigate away from that option.

To prevent errors like this in future, I added two debug error warnings, notifying me or any other developer if the assigned first selection for a window is disabled.