Tweaked options menu selectable colours, for improved readability when using a controller
I noticed that the colour changes appearing on the slider, toggle and dropdown prefabs I made had incredibly dull and similar colours, making it extremely difficult to tell which value the player was currently selecting (unless the player was using a mouse and could see the cursor). I tweaked the colour values to be more prominent, specifically by making them the same as for the other menu buttons, which also slightly improved visual consistency.