Fixed bug where pressing ‘Apply’ or ‘Revert’ button in options menu causes the player’s current selection to be stuck, when using a controller
This was a similar problem to one I had shortly before, where I would accidentally assign disabled option as the first selection when entering a menu, causing the player’s selection to be stuck unless they were using a pointer-based interface. In this case, it happened because pressing the Apply or Revert buttons would run the OptionsMenu’s Refresh() function, disabling the buttons’ interactability until the player changed an option again.
I fixed this with simple code that obtains the first selected option in the attached MenuWindow, and assigning it as the currently selected option just like how it does when entering a menu. Now, when the player presses ‘Apply’ or ‘Revert’, their selection is automatically switched to the first option.