Reworked control options menu code to show composite inputs and allow for easier implementation of rebinding code
The problem with the previous options menu code was that there was no way to show multiple bindings. The other problem was that I was having difficulty coming up with a good, modular way to implement button rebinding code. Unable to come up with a good solution, I decided I had to rework a lot of the functionality.
The code to auto-generate a list of titles and button prompts (with attached buttons to press to start rebinding a particular function) is still there, but some tweaks have been made. For each action, a title will be instantiated, then for each possible binding, an object is spawned featuring a ‘BindingOption’ class. This contains text and a visual prompt for the correct binding, as well as a reference to an InputAction and an array index, allowing it to reference and alter a specific binding within an InputAction class. This will allow me to spawn as many of these as are necessary for every possible binding of an action, and is much more modular.
There are some problems with this system, primarily that having all inputs in a single list makes it quite long and annoying to scroll through, and that making each option slimmer has made the icons quite small. But overall, I’m pleased with it, as it will now be far easier to implement button rebinding code, as I feel my core architecture is now a lot better suited to expandability.