Improved display names for input bindings
I added some extra formatting to the control binding names when they appear in the Control Options menu.
Bindings will now show the input type they are a part of (e.g. keyboard/mouse vs. gamepad). If they are part of a composite, the specific composite axis will be shown. The names are also formatted to add capital letters and remove awkward punctuation characters that existed for identification purposes in the rest of the input code. However, I have noticed a strange bug where some spaces won’t be added properly, which I will need to fix later.
I also had to rework similar formatting code in the GUIButtonPrompt class, by turning it into a public static function so it could be used by other functions in the game.