Temporarily removed/disabled all rebinding functionality to fix strange branch bugs
I noticed an unusual bug causing several of my prefabs to be declared missing, even when opening them would reveal them to be fine. After switching to a different branch, I noticed the problem got worse, and theorised it might be due to something going wrong when uploading to different branches. To fix this, I decided to either push or discard all changes in both my current branches, and merge everything together as a sort of reset.
Since I have a lot of half-finished rebinding code, I copied it into separate text documents and removed it from the scripts themselves. This way, I can ensure that only the functioning code is there and doesn’t cause any extra issues when combined with the rest of the scripts.