Progress on button rebinding
The player can click on a particular BindingOption’s button so the game starts checking for a new input (by enabling an auto-generated InputAction that registers any key press). When the player presses a particular input, its path will be formatted to a generic variant (so even if the player presses a button on a particular gamepad the input will be generic and work with any kind of gamepad), and checked against the current binding path. If the path is the same, or already taken by a different binding, or is incompatible with the control type of the previous binding, nothing will happen and an appropriate error message will be shown to the player. If the new path is compatible, it will be saved to a string within the BindingOption, so that if the player reverts their changes the original input layout is not affected.

The system is not complete, however. For example, input changes cannot actually be applied yet. This is because many bugs still plague the system. Most notably, sometimes trying to assign a certain binding path will return the path currently used by the binding being reassigned, even when the binding itself has not been changed. Additionally, certain buttons that also function as axes, such as D-pad buttons on a controller, simply won’t register.

Even though this system doesn’t fully work yet, I am relatively pleased with what I have so far.