Fixed bug causing weapon HUD to show incorrectly if the player is not holding a weapon
If the player did not have any weapons equipped, my HeadsUpDisplay script would cause a variety of NullReferenceException errors when trying to display weapon info. I fixed this with a simple check that only runs those functions if the player is currently holding a weapon, and disables the weapon HUD elements otherwise.