Custom state machine system
I created a custom state machine system that I will use to control the enemy’s AI routines. This consists of 2 classes and a struct.
The simplest but most important class, StateMachine.State, is an abstract class with five abstract functions – Enter(), Exit(), Update(), FixedUpdate() and LateUpdate(). The first will play once when the state is switched to, the second will play when it is switched away from, and the remaining three will run during the respective identically named MonoBehaviour update loops. As this class is abstract, I can easily create inherited classes that perform whatever actions I want. I created one test class, named StateMachine.TestState, which simply prints a message for the current state.
The struct is named Transition, and holds a reference to a ‘from’ State, a ‘to’ State, and a Func<bool> representing a transition condition. Since a bool is a value type and cannot be referenced without simply producing an identical value, using Func<bool> gives the Transition a reference so it can produce an accurate true/false value at any time.
The last class, StateMachine, is a MonoBehaviour that contains a list of State classes and Transition structs. This class is responsible for running the loops in the current State, and checking valid Transition structs to see if the state needs to be switched. If a transition does not have a specified ‘from’ State, this means it can be accessed from any other state.
This state machine is somewhat primitive (for example, it has no GUI so states and transitions must be generated and added in Awake(), but I am still very pleased with how it’s turned out so far. I feel the core elements are very solid and it will be easy to add more functionality without breaking any existing elements.